Allows disable cas even during asteroid avoiding
On every goto not to line 1 simple calculation that causes goto1 if :cas is disabled. Also f is set to 0 when :cas disabled.
Added this mod when accidently turned on cas on station pointing twards it. There where no way to stop it until it pointed large station almost around me.